Construct URI and call exchangeratesapi.io
Programmatically construct the URI used to call exchangeratesapi.io.
This allows a query parameter - currency - that can be used to specify
which currency should be analysed. Also ensure that data is retrieved
from the last week from runtime.

Take the response and unmarshal into a map[string]interface ready for
further processing.