Convert API response to array of Rate structs
Give some structure to the response obtained from exchangeratesapi.io.
Because the dates for each currency valuation are essentially a dynamic
key, the response has to be parsed into a nicer format.

This will make doing the computation for recommending whether to buy or
not buy currency later on easier.